Module Octez_smart_rollup

module Address : sig ... end

module Commitment : sig ... end

module Dal : sig ... end

module Dictionary_encoding : sig ... end

module Error : sig ... end

module Game : sig ... end

module Inbox : sig ... end

Smart rollup shared inbox representation, adapted from a subset of src/proto_alpha/lib_protocol/sc_rollup_inbox_repr.mli.

module Inbox_hash : sig ... end

module Installer_config : sig ... end

Handling of YAML configuration file used by the kernel installer.

module Kind : sig ... end

module L1_operation : sig ... end

module L2_message : sig ... end

module Merkelized_payload_hashes_hash : sig ... end

module Metadata : sig ... end

module Operation_kind : sig ... end

module Outbox_message : sig ... end

module Rollup_constants : sig ... end

Protocol agnostic representation for protocol constants.

module Rollup_node_services : sig ... end

This file defines the services of the rollup node that do not depend on the protocol. A top-level directory is built for them in the rollup node library.

module Sc_rollup_block : sig ... end

module State_hash : sig ... end

module Versioned_data : sig ... end

Values are versioned so that they can be modified in future iterations of the protocol.