Module Dal.Slot_history

module V1 : sig ... end

V1 are serialized slot histories as they appear first in Mumbai. Introduce V2 if encoding in future protocol changes.

include Versioned_data.S with type t = V1.t

type versioned

The type of versioned values, including the one that is currently used and the previous ones.

val versioned_encoding : versioned Tezos_base.TzPervasives.Data_encoding.t

Encoding for versioned values, corresponds to the one defined by the protocol.

val of_versioned : versioned -> V1.t

Conversion to the currently used representation. In practice this function must provide ways to translate values of older versions to the latest one.

val to_versioned : V1.t -> versioned

Wrapping a current value as a versioned one.

include module type of V1 with type t = V1.t

V1 are serialized slot histories as they appear first in Mumbai. Introduce V2 if encoding in future protocol changes.

type t = V1.t

val encoding : t Tezos_base.TzPervasives.Data_encoding.t