Module Octez_smart_rollup.Installer_config

Handling of YAML configuration file used by the kernel installer.

A configuration has the form: ``` instructions:

set: value: <hexadecimal value> to: <string> ... ```

type Tezos_base.TzPervasives.error +=

| Installer_config_yaml_error of string

| Installer_config_invalid

| Installer_config_invalid_instruction of int

type instr =

| Set of {

value : string;

to_ : string;

}

type t = instr list

val make : key:string -> value:string -> instr

val pp : Stdlib.Format.formatter -> t -> unit

val instr_encoding : instr Tezos_base.TzPervasives.Data_encoding.t

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

val parse_yaml : string -> t Tezos_base.TzPervasives.tzresult

parse_yaml content parses content as a YAML representing the configuration of the kernel installer, and returns the corresponding set of instructions.

val generate_yaml : t -> Yaml.yaml Tezos_base.TzPervasives.tzresult

generate_yaml instrs generates the YAML representation of instrs.

val emit_yaml : t -> string Tezos_base.TzPervasives.tzresult

emit_yaml instrs generates the YAML representation of instrs in textual format.

val to_file : t -> output:string -> unit Tezos_base.TzPervasives.tzresult Lwt.t

to_file instrs ~output writes the instructions into a file, either in JSON or YAML depending on the extension of output.