Module Octez_smart_rollup.L2_message

type t

Type of L2 messages.

val make : ?order:Z.t -> unique:bool -> string -> t

make ~unique message constructs a message with content message. If unique = true the returned value will be unique (multiple identical calls to make will return different values) because it will be given a unique id. Using unique = false, on the contrary, makes the call idempotent.

val content : t -> string

content message returns the string content of message, i.e. content (make s) = s.

module Id : Tezos_crypto.Intfs.HASH

Hash with b58check encoding scmsg(55), for ids of L2 messages.

type id = Id.t

Alias for message id

Serialization

val content_encoding : string Tezos_base.TzPervasives.Data_encoding.t

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

val id : t -> Id.t

val counter : t -> Z.t

val order : t -> Z.t option

val compare : t -> t -> int