Module Octez_smart_rollup.Operation_kind

type t =

| Publish

| Add_messages

| Cement

| Timeout

| Refute

| Recover

| Execute_outbox_message

| Publish_dal_commitment

The kind of operations that can be injected by the rollup node.

val all : t list

List of possible operations kind for operator specialization.

module Map : Tezos_base.TzPervasives.Map.S with type key = t

val to_string : t -> string

to_string o returns a string representation of operation_kind o.

val of_string : string -> t option

of_string s parses an operation kind from the given string s.

val of_string_exn : string -> t

of_string_exn s parses an operation kind from the given string s.

val encoding : t Tezos_base.TzPervasives.Data_encoding.t

val map_encoding : (t -> 'value Tezos_base.TzPervasives.Data_encoding.t) -> 'value Map.t Tezos_base.TzPervasives.Data_encoding.t

val compare_priority : t -> t -> int

compare_priority Comparison over tag for order of importance. Order is given by the list all. it's the following:

Timeout < Refute < Publish; Publish_dal_commitment; Cement < Recover < Add_messages < Execute_outbox_message.