Module Octez_smart_rollup.Outbox_message

type transaction_summary = {

destination : string;

entrypoint : string;

parameters : string Tezos_micheline.Micheline.canonical;

parameters_ty : string Tezos_micheline.Micheline.canonical option;

}

type summary =

| Whitelist_update of Tezos_crypto.Signature.Public_key_hash.t list option

| Transaction_batch of transaction_summary list

Simplified view of outbox messages for filtering and logging. This type must cover the values from all protocols.

val summary_encoding : summary Tezos_base.TzPervasives.Data_encoding.t

val pp_summary : Stdlib.Format.formatter -> summary -> unit