Module Octez_smart_rollup.Rollup_node_services

This file defines the services of the rollup node that do not depend on the protocol. A top-level directory is built for them in the rollup node library.

Protocol specific services are for RPCs under /global/block/ and are defined in the protocol specific libraries in src/proto_*/lib_sc_rollup_layer2/sc_rollup_services.ml.

We distinguish RPC endpoints served by the rollup node into global and local. The difference between the two lies in whether the responses given by different rollup nodes in the same state (see below for an exact definition) must be the same (in the case of global endpoints) or can differ (in the case of local endpoints).

More formally, two rollup nodes are in the same quiescent state if they are subscribed to the same rollup address, and have processed the same set of heads from the layer1. We only consider quiescent states, that is those where rollup nodes are not actively processing a head received from layer1.

Examples of global endpoints are current_tezos_head and last_stored_commitment, as the responses returned by these endpoints is expected to be consistent across rollup nodes in the same state.

An example of local endpoint is last_published_commitment, as two rollup nodes in the same state may either publish or not publish a commitment, according to whether its inbox level is below the inbox level of the last cemented commitment at the time they tried to publish the commitment. See below for a more detailed explanation.

type message_status =

| Unknown

| Pending_batch

| Pending_injection of L1_operation.t

| Injected of {

op : L1_operation.t;

oph : Tezos_base.TzPervasives.Operation_hash.t;

op_index : int;

}

| Included of {

op : L1_operation.t;

oph : Tezos_base.TzPervasives.Operation_hash.t;

op_index : int;

l1_block : Tezos_base.TzPervasives.Block_hash.t;

l1_level : int32;

finalized : bool;

cemented : bool;

}

| Committed of {

op : L1_operation.t;

oph : Tezos_base.TzPervasives.Operation_hash.t;

op_index : int;

l1_block : Tezos_base.TzPervasives.Block_hash.t;

l1_level : int32;

finalized : bool;

cemented : bool;

commitment : Commitment.t;

commitment_hash : Octez_smart_rollup.Commitment.Hash.t;

first_published_at_level : int32;

published_at_level : int32;

}

type gc_info = {

first_available_level : int32;

last_gc_started_at : int32 option;

last_context_split_level : int32 option;

last_successful_gc_target : int32 option;

}

type sync_result =

| Synchronized

| Synchronizing of {

processed_level : int32;

l1_head_level : int32;

percentage_done : float;

}

type l1_health = {

connection : [ `Connected | `Disconnected | `Reconnecting ];

blocks_late : int32;

last_seen_head : (Tezos_base.TzPervasives.Block_hash.t * int32 * Tezos_base.Time.Protocol.t) option;

}

type health = {

healthy : bool;

degraded : bool;

l1 : l1_health;

active_workers : (string * [ `Running | `Crashed of exn ]) list;

}

type version = {

version : string;

store_version : string;

context_version : string;

}

type injector_operation = {

op : L1_operation.t;

errors : int;

last_error : Tezos_base.TzPervasives.tztrace option;

}

type committed_status =

| Uncommitted of {

commitment_level : int32;
(*
Level at which the following commitment will be produced.
*)

estimated_commitment_time : Tezos_base.Time.System.t;
(*
Estimated time at which the commitment will be published.
*)

estimated_cementation_time : Tezos_base.Time.System.t;
(*
Estimated time at which the commitment will be cemented.
*)

}

| Committed_ of {

commitment_hash : Octez_smart_rollup.Commitment.Hash.t;
(*
Hash of following commitment.
*)

commitment_level : int32;
(*
Level of following commitment.
*)

first_published_level : int32;
(*
Level at which commitment was first published.
*)

published_level : int32 option;
(*
Level at which commitment was published by operator (may be None if e.g. rollup node does not publish).
*)

estimated_cementation_time : Tezos_base.Time.System.t;
(*
Estimated time at which the commitment will be cemented.
*)

}

| Cemented of {

commitment_hash : Octez_smart_rollup.Commitment.Hash.t;
(*
Hash of following commitment.
*)

commitment_level : int32;
(*
Level of following commitment.
*)

first_published_level : int32;
(*
Level at which commitment was first published.
*)

published_level : int32 option;
(*
Level at which commitment was published by operator (may be None if e.g. rollup node does not publish).
*)

blocks_since_cemented : int32;
(*
Number of blocks since commitment was cemented
*)

}

module Encodings : sig ... end

module Arg : sig ... end

module Query : sig ... end

module type PREFIX = sig ... end

module Make_services (P : PREFIX) : sig ... end

module Global : sig ... end

module Local : sig ... end

module Root : sig ... end

module Admin : sig ... end