Module type Versioned_data.S

type t

The type of values that is currently used.

type versioned

The type of versioned values, including the one that is currently used and the previous ones.

val versioned_encoding : versioned Tezos_base.TzPervasives.Data_encoding.t

Encoding for versioned values, corresponds to the one defined by the protocol.

val of_versioned : versioned -> t

Conversion to the currently used representation. In practice this function must provide ways to translate values of older versions to the latest one.

val to_versioned : t -> versioned

Wrapping a current value as a versioned one.