Module Octez_smart_rollup_node_store

module Sql_store : sig ... end

module Store : sig ... end

The store module is a pointer to the latest version. However, in order to run long migrations in the background, this module allows to have a store which is a hybrid version of the last two versions.

module Store_migration : sig ... end

module Store_sig : sig ... end

module Store_v0 : sig ... end

This version of the store is used for the rollup nodes for protocol Mumbai, i.e. = 16.

module Store_v1 : sig ... end

This version of the store is used for the rollup nodes for protocols for and after Nairobi, i.e. >= 17.

module Store_v2 : sig ... end

This version of the store is used for the rollup nodes for protocols for and after Nairobi, i.e. >= 17.

module Store_v3 : sig ... end

module Store_v4 : sig ... end

module Store_v5 : sig ... end

module Store_version : sig ... end