Module Sql_store.Commitments

Storage containing commitments and corresponding commitment hashes that the rollup node has knowledge of.

val store : ?conn:Octez_sqlite.Sqlite.conn -> rw -> Octez_smart_rollup.Commitment.t -> Octez_smart_rollup.Commitment.Hash.t Tezos_base.TzPervasives.tzresult Lwt.t

store ?conn s commitment stores a new commitment associated to its hash and returns it.

val find : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> Octez_smart_rollup.Commitment.Hash.t -> Octez_smart_rollup.Commitment.t option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve a commitment by its hash.

val find_lcc : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> unit -> Octez_smart_rollup.Commitment.t option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve a commitment corresponding to the LCC, the last cemented commitment.

val find_lpc : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> Octez_smart_rollup.Commitment.t option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve the last commitment published by the rollup node (as known from the handled L1 blocks).