Module Sql_store.Commitments_published_at_levels

Storage mapping commitment hashes to the level when they were published by the rollup node.

type publication_levels = {

first_published_at_level : int32;
(*
The level at which this commitment was first published.
*)

published_at_level : int32 option;
(*
The level at which we published this commitment. If first_published_at_level <> published_at_level it means that the commitment is republished.
*)

}

val register : ?conn:Octez_sqlite.Sqlite.conn -> rw -> Octez_smart_rollup.Commitment.Hash.t -> publication_levels -> unit Tezos_base.TzPervasives.tzresult Lwt.t

register ?conn s hash levels stores the publication levels for commitment whose hash is hash.

val get : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> Octez_smart_rollup.Commitment.Hash.t -> publication_levels option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve the publication levels for a commitment by its hash.

val get_first_published_level : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> Octez_smart_rollup.Commitment.Hash.t -> int32 option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve the level at which a commitment was first published.