Module Sql_store.Dal_slots_headers

Published slot headers per block hash, stored as a list of bindings from Dal_slot_index.t to Dal.Slot.t. The encoding function converts this list into a Dal.Slot_index.t-indexed map.

val store : ?conn:Octez_sqlite.Sqlite.conn -> rw -> Tezos_base.TzPervasives.Block_hash.t -> Octez_smart_rollup.Dal.Slot_header.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

store ?conn s block slot_header stores the slot header slot_header as being published in the L1 block whose hash is block.

val find_slot_header : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> Tezos_base.TzPervasives.Block_hash.t -> slot_index:Octez_smart_rollup.Dal.Slot_index.t -> Octez_smart_rollup.Dal.Slot_header.t option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve a published slot header by its publication block and slot index.

val list_slot_headers : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> Tezos_base.TzPervasives.Block_hash.t -> Octez_smart_rollup.Dal.Slot_header.t list Tezos_base.TzPervasives.tzresult Lwt.t

List all slot headers published in a block.

val list_slot_indexes : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> Tezos_base.TzPervasives.Block_hash.t -> Octez_smart_rollup.Dal.Slot_index.t list Tezos_base.TzPervasives.tzresult Lwt.t

List all indexes of slot headers published in a block.