Module Sql_store.Dal_slots_statuses

Dal_slots_statuses is used to store the attestation status of DAL slots. The values of this storage module have type ``Confirmed | `Unconfirmed`, depending on whether the content of the slot has been attested on L1 or not. If an entry is not present for a (block_hash, slot_index), this means that the corresponding block is not processed yet.

val store : ?conn:Octez_sqlite.Sqlite.conn -> rw -> Tezos_base.TzPervasives.Block_hash.t -> Octez_smart_rollup.Dal.Slot_index.t -> [ `Confirmed | `Unconfirmed ] -> unit Tezos_base.TzPervasives.tzresult Lwt.t

store ?conn s block slot_index status store the attestation status of slot published in block and index slot_index.

val find_slot_status : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> Tezos_base.TzPervasives.Block_hash.t -> slot_index:Octez_smart_rollup.Dal.Slot_index.t -> [ `Confirmed | `Unconfirmed ] option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve the attestation status of a DAL slot.

val list_slot_statuses : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> Tezos_base.TzPervasives.Block_hash.t -> (Octez_smart_rollup.Dal.Slot_index.t * [ `Confirmed | `Unconfirmed ]) list Tezos_base.TzPervasives.tzresult Lwt.t

List attestation statuses of all slots published in a given L1 block.