Module Sql_store.Inboxes

Aggregated collection of messages from the L1 inbox

val store : ?conn:Octez_sqlite.Sqlite.conn -> rw -> Octez_smart_rollup.Inbox.t -> Octez_smart_rollup.Inbox_hash.t Tezos_base.TzPervasives.tzresult Lwt.t

store ?conn s inbox stores the inbox, associated to its hash and returns it.

val find : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> Octez_smart_rollup.Inbox_hash.t -> Octez_smart_rollup.Inbox.t option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve an inbox by its hash.

val find_by_block_hash : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> Tezos_base.TzPervasives.Block_hash.t -> Octez_smart_rollup.Inbox.t option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve an inbox by its the block hash in which it appeared.