Module Sql_store.L2_levels

Storage for associating levels to block hashes

val store : ?conn:Octez_sqlite.Sqlite.conn -> rw -> int32 -> Tezos_base.TzPervasives.Block_hash.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

store ?conn s level block associates the block hash block to level.

val find : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> int32 -> Tezos_base.TzPervasives.Block_hash.t option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve the block hash associated to a given level.