Module Sql_store.Outbox_messages

Storage for persisting outbox messages.

val pending : ?conn:Octez_sqlite.Sqlite.conn -> _ t -> min_level:int32 -> max_level:int32 -> (int32 * int list) list Tezos_base.TzPervasives.tzresult Lwt.t

pending ?conn s ~min_level ~max_level returns all pending (i.e. non executed) outbox messages between outbox levels min_level and max_level. The result is given as a list of pairs whose first component is the outbox level and the second is the message indexes list.

val register_outbox_messages : ?conn:Octez_sqlite.Sqlite.conn -> rw -> outbox_level:int32 -> indexes:Tezos_base.Bitset.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Register outbox messages for a given outbox level by its indexes.

val set_outbox_message_executed : ?conn:Octez_sqlite.Sqlite.conn -> rw -> outbox_level:int32 -> index:int -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Register an outbox message as executed by its outbox level and its index in the outbox.