Module Octez_smart_rollup_node_store.Store

The store module is a pointer to the latest version. However, in order to run long migrations in the background, this module allows to have a store which is a hybrid version of the last two versions.

include module type of struct include Store_v5 end

val version : Store_version.t

include module type of struct include Sql_store end

val sqlite_file_name : string

Name of SQLite file in data directory.

val extra_sqlite_files : string list

Other files that make up the store.

val init : 'a Tezos_layer2_store.Store_sigs.mode -> data_dir:string -> 'a Sql_store.t Tezos_base.TzPervasives.tzresult Lwt.t

init mode ~data_dir initializes the store and returns it.

val export_store : data_dir:string -> output_db_file:string -> unit Tezos_base.TzPervasives.tzresult Lwt.t

export_store ~data_dir ~output_db_file exports the store database with data from the data_dir into the output_db_file. This function also removes data that is specific to the operator. This function is meant to be used to produce snapshots.

val with_transaction : _ Sql_store.t -> (Octez_sqlite.Sqlite.conn -> 'a Tezos_base.TzPervasives.tzresult Lwt.t) -> 'a Tezos_base.TzPervasives.tzresult Lwt.t

with_transaction store f executes f with a single connection to the database in a transaction. I.e., if f returns an error or raises an exception, the store will not be modified.

Modules for storing data in the store

Each module maps to a table in the database. Every read and write function has two parameters: store and an optional ?conn.

When ~conn is provided, the database connection is used directly to execute the queries. If not, a new (or old) connection is taken from the pool of store.

type 'a hybrid = {

v4 : Store_v4.ro;
(*
A read-only old V4 store.
*)

v5 : 'a Store_v5.t;
(*
A read-write V5 store to which we migrate data and which we use to store new data.
*)

migration_done : bool Stdlib.ref;
(*
A flag to indicate when the migration has terminated. When this is the case we only query the v5 store.
*)

}

The hybrid store.

type nonrec 'a t =

| Hybrid of 'a hybrid
(*
Either a hybrid store when a bacground migration is running.
*)

| Normal of 'a {t}1
(*
Or a "normal" store, for the nominal case.
*)

The type of stores.

type rw = Tezos_layer2_store.Store_sigs.rw t

type ro = Tezos_layer2_store.Store_sigs.ro t

val with_rw : 'a Tezos_layer2_store.Store_sigs.mode -> (unit -> rw Tezos_base.TzPervasives.tzresult Lwt.t) -> 'a t Tezos_base.TzPervasives.tzresult Lwt.t

val init_with_migration : 'a Tezos_layer2_store.Store_sigs.mode -> Octez_smart_rollup.Metadata.t -> data_dir:string -> 'a t Tezos_base.TzPervasives.tzresult Lwt.t

Initialize a store and run the corresponding migrations if needed. For the v4 -> v5 transition, the migration is run in the background.

val close : 'a t -> unit Lwt.t

val readonly : 'a t -> Tezos_layer2_store.Store_sigs.ro t

val gc : Tezos_layer2_store.Store_sigs.rw t -> level:int32 -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val read_from : 'a t -> 'a t

val write_to : 'a t -> 'a {t}1

val (@?~~>) : ('a option, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t -> ('a option, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t -> ('a option, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

x @?~~> y executes x first and then y if x resolves to None (where x is an access to the V5 store and y to the V4 store).

module Commitments : sig ... end

module Commitments_published_at_levels : sig ... end

module Inboxes : sig ... end

module Messages : sig ... end

module Outbox_messages : sig ... end

module Protocols : sig ... end

module Dal_slots_headers : sig ... end

module Dal_slots_statuses : sig ... end

module L2_levels : sig ... end

module L2_blocks : sig ... end

module State : sig ... end