Module Store.Commitments_published_at_levels

Storage mapping commitment hashes to the level when they were published by the rollup node.

type publication_levels = {

first_published_at_level : int32;
(*
The level at which this commitment was first published.
*)

published_at_level : int32 option;
(*
The level at which we published this commitment. If first_published_at_level <> published_at_level it means that the commitment is republished.
*)

}

val register : Tezos_layer2_store.Store_sigs.rw t -> Octez_smart_rollup.Commitment.Hash.t -> publication_levels -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val get : 'a t -> Octez_smart_rollup.Commitment.Hash.t -> publication_levels option Tezos_base.TzPervasives.tzresult Lwt.t

val get_first_published_level : 'a t -> Octez_smart_rollup.Commitment.Hash.t -> (int32 option, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t