Module Octez_smart_rollup_node_store.Store_migration

val maybe_run_migration : Octez_smart_rollup.Metadata.t -> Store_version.t -> data_dir:string -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Migrate store located in rollup node data directory data_dir if needed. If there is no possible migration path registered to go from the current version to the last Store.version, this function resolves with an error.

Version specific standalone migration functions

module V5_sqlite_migrations : sig ... end