Module Store_v0.Irmin_store

include sig ... end

type +'a t

val name : string

val make_key_path : string list -> string -> string list

val load : 'a Tezos_layer2_store__Store_sigs.mode -> string -> 'a t Tezos_base.TzPervasives.tzresult Lwt.t

val flush : [> `Write ] t -> unit Tezos_base.TzPervasives.tzresult

val close : 'a t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val set : [> `Write ] t -> string list -> bytes -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val get : [> `Read ] t -> string list -> bytes Tezos_base.TzPervasives.tzresult Lwt.t

val mem : [> `Read ] t -> string list -> bool Tezos_base.TzPervasives.tzresult Lwt.t

val find : [> `Read ] t -> string list -> bytes option Tezos_base.TzPervasives.tzresult Lwt.t

val path_to_string : string list -> string

val readonly : [> `Read ] t -> [ `Read ] t

module Raw_irmin : sig ... end

include Tezos_layer2_store.Store_sigs.Store with type 'a t := 'a t

module Make_updatable_map (S : Tezos_layer2_store.Store_sigs.STORAGE_INFO) (K : Tezos_layer2_store.Store_sigs.KEY) (V : Tezos_layer2_store.Store_sigs.VALUE) : Tezos_layer2_store.Store_sigs.Map with type 'a store = 'a t and type key = K.key and type value = V.value

Make_updatable_map(S)(K)(V) constructs a Map which can be persisted on store. The module S defines storage-dependent information about how the map will be saved on and retrieved from the store (for example, it defines the map location in the store). The module K defines the information related to keys of the map, and the module V contains information about how values will be stored to and retrieved from the store. The resulting map allows to update the contents of an existing value for a key.

module Make_append_only_map (S : Tezos_layer2_store.Store_sigs.STORAGE_INFO) (K : Tezos_layer2_store.Store_sigs.KEY) (V : Tezos_layer2_store.Store_sigs.VALUE) : Tezos_layer2_store.Store_sigs.Append_only_map with type 'a store = 'a t and type key = K.key and type value = V.value

Make_append_only_map(S)(K)(V) constructs an Append_only_map which can be persisted on store. The module S defines storage-dependent information about how the map will be saved on and retrieved from the store (for example, it defines the map location in the store). The module K contains information related to keys of the map, and the module V contains information about how values will be stored to and retrieved from the store. The resulting map forbids updating the contents of an existing value with a new value, different from the previous one.

module Make_mutable_value (S : Tezos_layer2_store.Store_sigs.STORAGE_INFO) (V : Tezos_layer2_store.Store_sigs.VALUE) : Tezos_layer2_store.Store_sigs.Mutable_value with type 'a store = 'a t and type value = V.value

Make_mutable_value(S)(V) constructs a Mutable_value for persisting a mutable value in a store. The module parameter S defines the location of the mutable value in the store, and the module parameter V contains information about the type of values that the constructed module will persist in the underlying store.

module Make_nested_map (S : Tezos_layer2_store.Store_sigs.STORAGE_INFO) (K1 : Tezos_layer2_store.Store_sigs.KEY) (K2 : Tezos_layer2_store.Store_sigs.COMPARABLE_KEY) (V : Tezos_layer2_store.Store_sigs.VALUE) : Tezos_layer2_store.Store_sigs.Nested_map with type 'a store = 'a t and type primary_key = K1.key and type secondary_key = K2.key and type value = V.value

Make_nested_map(S)(K1)(K2)(V) constructs a Nested_map module using module parameter S to define where the map is going to be persisted on store, K1 and K2 to define the primary and secondary key, respectively, and V to define the values of the resulting Nested_map.