Module Store_v2.Gc_levels

Data related to the effects of garbage collection.

type levels = {

last_gc_level : int32;
(*
Records the last level at which GC was called.
*)

first_available_level : int32;
(*
Records the first level for which data is guaranteed to be stored. Data for all previous levels might have been removed.
*)

}

include Tezos_layer2_store.Indexed_store.SINGLETON_STORE with type value = levels

type +'a t

The type of the singleton store.

type value = levels

The type of values stored in this singleton store.

val load : path:string -> 'a Tezos_layer2_store.Store_sigs.mode -> 'a t Tezos_base.TzPervasives.tzresult Lwt.t

Load (or initializes) a singleton store in the file path.

val read : [> `Read ] t -> value option Tezos_base.TzPervasives.tzresult Lwt.t

Reads the current value from the disk. Returns None if the file does not exist.

val write : [> `Write ] t -> value -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Write the value to disk.

val delete : [> `Write ] t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Deletes the value from the disk.

val readonly : [> `Read ] t -> [ `Read ] t

readonly t returns a read only version of the store t.