Module Store_v2.Protocols

type level =

| First_known of int32

| Activation_level of int32

type proto_info = {

level : level;
(*
The level at which we have seen the protocol for the first time, either because we saw its activation or because the first block we saw (at the origination of the rollup) was from this protocol.
*)

proto_level : int;
(*
The protocol level, i.e. its number in the sequence of protocol activations on the chain.
*)

protocol : Tezos_base.TzPervasives.Protocol_hash.t;
(*
The protocol this information concerns.
*)

}

Each element of this type represents information we have about a Tezos protocol regarding its activation.

val proto_info_encoding : proto_info Tezos_base.TzPervasives.Data_encoding.t

include Tezos_layer2_store.Indexed_store.SINGLETON_STORE with type value = proto_info list

type +'a t

The type of the singleton store.

type value = proto_info list

The type of values stored in this singleton store.

val load : path:string -> 'a Tezos_layer2_store.Store_sigs.mode -> 'a t Tezos_base.TzPervasives.tzresult Lwt.t

Load (or initializes) a singleton store in the file path.

val read : [> `Read ] t -> value option Tezos_base.TzPervasives.tzresult Lwt.t

Reads the current value from the disk. Returns None if the file does not exist.

val write : [> `Write ] t -> value -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Write the value to disk.

val delete : [> `Write ] t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Deletes the value from the disk.

val readonly : [> `Read ] t -> [ `Read ] t

readonly t returns a read only version of the store t.