Module Octez_smart_rollup_node_store.Store_v5

val version : Store_version.t

include module type of struct include Sql_store end

type 'a t = 'a Sql_store.t

Type of store (a handler to the underlying database).

type rw = Tezos_layer2_store.Store_sigs.rw t

Read/write store t.

type ro = Tezos_layer2_store.Store_sigs.ro t

Read only store t.

val sqlite_file_name : string

Name of SQLite file in data directory.

val extra_sqlite_files : string list

Other files that make up the store.

val init : 'a Tezos_layer2_store.Store_sigs.mode -> data_dir:string -> 'a t Tezos_base.TzPervasives.tzresult Lwt.t

init mode ~data_dir initializes the store and returns it.

val close : _ t -> unit Lwt.t

Close the store by freeing all resources and closing database connections.

val readonly : _ t -> ro

Returns a read-only version of the store.

val gc : rw -> level:int32 -> unit Tezos_base.TzPervasives.tzresult Lwt.t

gc store ~level garbage collects data that relate to levels below level (by removing information from the database).

val export_store : data_dir:string -> output_db_file:string -> unit Tezos_base.TzPervasives.tzresult Lwt.t

export_store ~data_dir ~output_db_file exports the store database with data from the data_dir into the output_db_file. This function also removes data that is specific to the operator. This function is meant to be used to produce snapshots.

val with_transaction : _ t -> (Octez_sqlite.Sqlite.conn -> 'a Tezos_base.TzPervasives.tzresult Lwt.t) -> 'a Tezos_base.TzPervasives.tzresult Lwt.t

with_transaction store f executes f with a single connection to the database in a transaction. I.e., if f returns an error or raises an exception, the store will not be modified.

Modules for storing data in the store

Each module maps to a table in the database. Every read and write function has two parameters: store and an optional ?conn.

When ~conn is provided, the database connection is used directly to execute the queries. If not, a new (or old) connection is taken from the pool of store.

module Commitments = Sql_store.Commitments

Storage containing commitments and corresponding commitment hashes that the rollup node has knowledge of.

module Commitments_published_at_levels = Sql_store.Commitments_published_at_levels

Storage mapping commitment hashes to the level when they were published by the rollup node.

module Inboxes = Sql_store.Inboxes

Aggregated collection of messages from the L1 inbox

module Messages = Sql_store.Messages

Storage for persisting messages downloaded from the L1 node.

module Outbox_messages = Sql_store.Outbox_messages

Storage for persisting outbox messages.

module Protocols = Sql_store.Protocols

Storage for protocol activation levels.

module Dal_slots_headers = Sql_store.Dal_slots_headers

Published slot headers per block hash, stored as a list of bindings from Dal_slot_index.t to Dal.Slot.t. The encoding function converts this list into a Dal.Slot_index.t-indexed map.

module Dal_slots_statuses = Sql_store.Dal_slots_statuses

Dal_slots_statuses is used to store the attestation status of DAL slots. The values of this storage module have type ``Confirmed | `Unconfirmed`, depending on whether the content of the slot has been attested on L1 or not. If an entry is not present for a (block_hash, slot_index), this means that the corresponding block is not processed yet.

module L2_levels = Sql_store.L2_levels

Storage for associating levels to block hashes

module L2_blocks = Sql_store.L2_blocks

Storage for L2 blocks contracted by the rollup node.

module State = Sql_store.State

Storage of single values.