Module Octez_smart_rollup_wasm_benchmark_lib.Data

How to collect and store data

type datum

container of the informations on a data point

val make_datum : string -> string -> string -> Z.t -> Measure.time -> datum

make_datum scenario section label ticks time

type benchmark

container of all data point informations collected during benchmark

val empty_benchmark : ?verbose:bool -> ?totals:bool -> ?irmin:bool -> unit -> benchmark

initialize en empty benchmark with options

verbose: ouput info during execution (besides csv data in the end)

totals: add to csv data the total time / tick number for each steps

irmin: add data points corresponding to irmin decoding / encoding

val init_scenario : string -> benchmark -> benchmark

init_scenario scenario_name benchmark inits an empty benchmark for a given scenario

val switch_section : string -> benchmark -> benchmark

switch_section section_name benchmark open a new section

val add_datum : string -> Z.t -> Measure.time -> benchmark -> benchmark

add_datum name ticks time benchmark

val add_tickless_datum : string -> Measure.time -> benchmark -> benchmark

add_tickless_datum label time benchmark adds a point of data for an action consuming no tick

val add_final_info : Measure.time -> Z.t -> benchmark -> benchmark

adds final info as a data point in the benchmark

module Csv : sig ... end

val pp_analysis : Stdlib.Format.formatter -> benchmark -> unit