Module Inputs.Kernels

val tx_kernel_dac : string

val computation_kernel : string

taken from unit test, reads input, execute without taking them into account

val unreachable_kernel : string

taken from unit tests, supposed to get stuck at the first execution