Module Octez_smart_rollup_wasm_benchmark_lib.Measure

type time

val time : (unit -> 'a Lwt.t) -> (time * 'a) Lwt.t

time f computes f () returns the time taken to compute and the result

val time_and_tick : ('a -> Z.t Lwt.t) -> ('a -> 'a Lwt.t) -> 'a -> (time * Z.t * 'a) Lwt.t

time_and_tick get_tick f a computes f a and returns the time taken, the number of ticks and the result

val zero : time

val to_seconds : time -> float