Module Octez_sqlite.Sqlite

type t

A handler to the database.

type conn

A direct connection to the database, allowing to interact with it.

Initialization and backup

val init : path:string -> perm:[ `Read_only | `Read_write ] -> (conn -> unit Tezos_error_monad.Error_monad.tzresult Lwt.t) -> t Tezos_error_monad.Error_monad.tzresult Lwt.t

init ~path ~perm migratrion returns a handler to the database located at path and executes the given migration code.

If sqlite_journal_mode is `Force mode, then the journal mode of the SQLite database is updated if necessary to match the requested configuration. With `Identity, the journal mode is left untouched.

If perm is `Read_only, then SQL requests requiring write access will fail. With `Read_write, they will succeed as expected.

val close : t -> unit Lwt.t

val vacuum : conn:conn -> output_db_file:string -> unit Tezos_error_monad.Error_monad.tzresult Lwt.t

Rebuild the database in output_db_file using the VACUUM sqlite command. This function is useful to backup the database.

val vacuum_self : conn:conn -> unit Tezos_error_monad.Error_monad.tzresult Lwt.t

Vacuums the database itself after removing lot of data VACUUM sqlite command.

Database connections

val use : t -> (conn -> 'a Tezos_error_monad.Error_monad.tzresult Lwt.t) -> 'a Tezos_error_monad.Error_monad.tzresult Lwt.t

use db k executes k with a fresh connection to db.

val with_transaction : conn -> (conn -> 'a Tezos_error_monad.Error_monad.tzresult Lwt.t) -> 'a Tezos_error_monad.Error_monad.tzresult Lwt.t

with_transaction conn k wraps the accesses to the database from conn made in the continuation k within a SQL transaction. If k fails, the transaction is rollbacked. Otherwise, the transaction is committed.

val assert_in_transaction : conn -> unit

assert_in_transaction conn raises an exception if a transaction has not been started with conn.

raises Assert_failure

val with_connection : conn -> ((module Caqti_lwt.CONNECTION) -> 'a) -> 'a

with_connection conn k allows to wraps atomic low level accesses to the database from conn. with_connection can be used in the continuation of with_transaction.

Database low level queries

module Db : sig ... end

Caqti convenience functions wrapped in the Tezos error monad. See Caqti_connection_sig.Convenience.