Module Tezos_layer2_irmin_context.Irmin_context_events

val starting_context_gc : Tezos_layer2_store.Smart_rollup_context_hash.t -> unit Lwt.t

starting_context_gc hash emits an event which indicates that a GC run was launched for hash.

val context_gc_already_launched : unit -> unit Lwt.t

context_gc_already_launched () emits an event which indicates that a GC launch was attempted but resulted in no action because a GC run is already in progress.

val ending_context_gc : (Ptime.span * Ptime.span) -> unit Lwt.t

ending_context_gc total_duration finalise_duration emits an event which indicates that a GC run has ended, providing its total duration and its finalisation duration.

val context_gc_failure : string -> unit Lwt.t

context_gc_failure err emits an event which indicates a GC failure.

val context_gc_launch_failure : string -> unit Lwt.t

context_gc_launch_failure err emits an event which indicates a GC launch error.