Module Indexed_store.Make_simple_indexed_file

Parameters

module _ : NAME

module K : INDEX_KEY

module V : sig ... end

Signature

include INDEXED_FILE with type key := K.t with type value := V.t with type header := V.Header.t

type +'a t

The type of indexed file store.

val mem : [> `Read ] t -> K.t -> bool Tezos_base.TzPervasives.tzresult Lwt.t

Returns true if the key has a value associated in the store.

val header : [> `Read ] t -> K.t -> V.Header.t option Tezos_base.TzPervasives.tzresult Lwt.t

Returns the header for a key if it exists in the store.

val read : [> `Read ] t -> K.t -> (V.t * V.Header.t) option Tezos_base.TzPervasives.tzresult Lwt.t

Read a full value and header from the indexed file store.

val load : path:string -> index_buffer_size:int -> cache_size:int -> 'a Store_sigs.mode -> 'a t Tezos_base.TzPervasives.tzresult Lwt.t

Loads a new or existing indexed file store in the directory path.

val close : _ t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Close the index and the file. One must call load again to read or write data in the store.

val readonly : [> `Read ] t -> [ `Read ] t

readonly t returns a read only version of the store t.

val gc : ?async:bool -> Store_sigs.rw t -> (K.t -> V.Header.t -> V.t -> bool Tezos_base.TzPervasives.tzresult Lwt.t) -> unit Tezos_base.TzPervasives.tzresult Lwt.t

gc ?async t filter garbage collects data stored in the store t by keeping only elements that satisfy the predicate filter. This call runs the GC asynchronously unless async is false. If a GC is already ongoing this new request is ignored and this call is a no-op.

val wait_gc_completion : 'a t -> unit Lwt.t

wait_gc_completion t returns a blocking thread if a GC run is currently ongoing.

val is_gc_finished : 'a t -> bool

is_gc_finished t returns true if there is no GC running.

val cancel_gc : 'a t -> bool Lwt.t

cancel_gc t stops the currently ongoing GC if any. It returns true if a GC was canceled.

val append : ?flush:bool -> [> `Write ] t -> key:K.t -> value:V.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Append a new binding to the indexed file store.