Module type Indexed_store.SINGLETON_STORE

A store for single updatable values. Values are stored in a file on disk and are kept in memory in a cache.

type +'a t

The type of the singleton store.

type value

The type of values stored in this singleton store.

val load : path:string -> 'a Store_sigs.mode -> 'a t Tezos_base.TzPervasives.tzresult Lwt.t

Load (or initializes) a singleton store in the file path.

val read : [> `Read ] t -> value option Tezos_base.TzPervasives.tzresult Lwt.t

Reads the current value from the disk. Returns None if the file does not exist.

val write : [> `Write ] t -> value -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Write the value to disk.

val delete : [> `Write ] t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Deletes the value from the disk.

val readonly : [> `Read ] t -> [ `Read ] t

readonly t returns a read only version of the store t.