Module Snapshot_utils.Make

Parameters

module Header : sig ... end

Signature

val create : reader -> writer -> Header.t -> files:(string * string) list -> dest:string -> unit

create reader writer header ~files ~dest creates a snapshot archive with the header header with the contents of files. Each element of files is a pair whose first component is the path of the file to include and the second component is the "relative" path it should be registered to in the snapshot archive The archive is produced in file dest.

val extract : reader -> writer -> (Header.t -> 'a Tezos_base.TzPervasives.tzresult Lwt.t) -> cancellable:bool -> display_progress:bool -> snapshot_file:string -> dest:string -> (Header.t * 'a) Tezos_base.TzPervasives.tzresult Lwt.t

extract reader writer check_header ~snapshot_file ~dest extracts the snapshot archive snapshot_file in the directory dest. Existing files in dest with the same names are overwritten. The header header read from the snapshot is checked with check_header before beginning extraction, and returned. If display_progress is set to true, a spinner is exposed to the user to amuse them while they wait. Set cancellable to true if you want to be able to use Lwt.cancel on the promise.

val compress : snapshot_file:string -> string

compress ~snapshot_file compresses the snapshot archive snapshot_file of the form "path/to/snapshot.uncompressed" to a new file "path/to/snapshot" whose path is returned. snapshot_file is removed upon successful compression.

val read_header : reader -> snapshot_file:string -> Header.t

read_header reader ~snapshot_file reads the header from the snapshot file without extracting it.