Module type Store_sigs.Mutable_value

Module type respresenting a Mutable_value that is persisted on store.

type +'a store

The type of the store that is used for persisting data on disk.

type value

The type of the values that will be persisted.

val path_key : path

Path in the irmin tree.

val set : [> `Write ] store -> value -> unit Tezos_base.TzPervasives.tzresult Lwt.t

set store value persists value for this Mutable_value on store.

val get : [> `Read ] store -> value Tezos_base.TzPervasives.tzresult Lwt.t

get store retrieves the value persisted in store for this Mutable_value. If the underlying storage backend fails to retrieve the contents of the mutable value by throwing an exception, then the exception is propagated by get store.

val find : [> `Read ] store -> value option Tezos_base.TzPervasives.tzresult Lwt.t

find store returns an optional value containing the value persisted in store for this Mutable_value. If no value is persisted for the Mutable_value, None is returned.