Parameter Make_nested_map.K2

type key

The type of a comparable key.

val name : string

name is a string to identify the value of keys.

val compare : key -> key -> int

compare key1 key2 compares key1 with key2.

val encoding : key Tezos_base.TzPervasives.Data_encoding.t

encoding is the encoding for values of type key.