Module Tezos_scoru_wasm_fast.Exec

val preload_kernel : hooks:Tezos_scoru_wasm.Hooks.t -> Tezos_scoru_wasm.Durable.t -> unit Lwt.t

preload_kernel ~hooks durable can be used to parse and load in the module cache the kernel from durable, as a way to speed-up further executions involving this kernel.

val compute : ?wasm_entrypoint:string -> hooks:Tezos_scoru_wasm.Hooks.t -> version:Tezos_scoru_wasm.Wasm_pvm_state.version -> reveal_builtins:Tezos_scoru_wasm.Builtins.reveals -> write_debug:Tezos_scoru_wasm.Builtins.write_debug -> Tezos_scoru_wasm.Durable.t -> Tezos_webassembly_interpreter.Eval.buffers -> Tezos_scoru_wasm.Durable.t Lwt.t

compute durable buffers execute wasm_entrypoint once. Defaults to Constants.wasm_entrypoint when omitted.