Module Tezos_scoru_wasm_fast.Funcs

type host_state = {

retrieve_mem : unit -> Tezos_wasmer.Memory.t;

buffers : Tezos_webassembly_interpreter.Eval.buffers;

mutable durable : Tezos_scoru_wasm.Durable.t;

}

Host state environment for Fast Execution

val make : version:Tezos_scoru_wasm.Wasm_pvm_state.version -> reveal_builtins:Tezos_scoru_wasm.Builtins.reveals -> write_debug:Tezos_scoru_wasm.Builtins.write_debug -> host_state -> (string * string * Tezos_wasmer.extern) list

make ~debug builtins host_state generates a list of host functions that can be imported by WebAssembly modules. If debug is true, an implementation for write_debug will be picked that prints to the standard output.