Module Tezos_scoru_wasm_fast.Module_cache

module Wasmer = Tezos_wasmer

val load_kernel : hooks:Tezos_scoru_wasm.Hooks.t -> Wasmer.Store.t -> Tezos_scoru_wasm.Durable.t -> Wasmer.Module.t Lwt.t

load_kernel store durable returns a module attached to the Wasmer store, corresponding to kernel stored in the durable storage. Module is cached, and not reparsed if it has already been instanciated.