Module Wasm_fast.Unsafe

val get_max_nb_ticks : tree -> Z.t Lwt.t

Retrieve the maximum number of ticks for the PVM from the state.

val set_max_nb_ticks : Z.t -> tree -> tree Lwt.t

Change the maximum number of ticks (per snapshot) of the WASM PVM. This is to be used only for tests or to increase the tick limit in a non-refutable setting.

val durable_set : key:string -> value:string -> tree -> tree Lwt.t

Set a value to a given key in the durable storage of tree. If there is an existing value, it is overwritten.