Module Memory.Internal_for_tests

val of_list : Unsigned.uint8 list -> t

of_list content creates a memory instance containing the bytes from content The content is expected to be a multiple of pages size (64KB)

val to_list : t -> Unsigned.uint8 list

to_list mem returns a list containing each byte of mem