Module Tezos_wasmer.Module

type t

WebAssembly module

type format =

| Text

| Binary

Textual or binary representation of WebAssembly

val create : Store.t -> format -> string -> t

create store format code parses a module in the given format.

val delete : t -> unit

delete module_ destroys a WebAssembly module. Make sure that module_ is not used after delete is called on it.